Jenkinsfile / Kubernetes Resources:
Updated the Jenkinsfile to add deploying to the Cluster:
    - AWS Load Balancer Controller
    - Kubernetes Service (ClusterIP)
    - Kubernetes Ingress

I added two yaml files for deploying the two kubernetes resources.
The AWS Load Balancer Controller required a lot of steps, and
all of them were done in the Jenkinsfile:
    - Create AWS IAM Policy (Might move this to Cloudformation)
    - Create Kubernetes Servcic (Might move this to kubernetes yaml)
    - Install AWS Load Balancer Controller using Helm

The AWS Load Balancer Controller should automatically create an ALB
in AWS that can be used to externally connect to the application
running in the Pod.

I am sure there are multiple errors with this.