Jenkinsfile - eks deployment and parameters.json:
Had to update the Jenkinsfile to wait until the
aws-load-balancer-controller was fully deployed
before moving on. I added logic for this, because
a simple sleep might cause issues down the road.

I might have to add more logic for the Jenkinsfile
to wait until kubernetes resources are deployed.
As kubectl returns true as long as it was able to
deploy the resources, it won't wait until they are
ready.

Lastly, I solved the problem of the BastionHost
UserData not passing in the parameters. This also
helped me to discover why the PUBLIC_IP parameter
kept failing when I tried to pass that in.

The Jenkins cred were getting passed in, but I
was using the parameters from parameters.json.
In parameters.json I was using the jenkins creds,
but those variable names weren't getting updated.
So, I added sed commands to actually pass the
values of the JenkinsCreds into parameters.json.