Jenkinsfile, IaC BH deployment, EKS deployment
Updated Jenkinsfile to try and fix the deployment
of the services and load balancers. Pretty sure
the problem was the service account and all the
configurations for the AWS Load Balancer Controller
was made in the kube-system namespace. Changed the
namespace to be 'web-app.'

Updated web-app-deployment to use the serviceaccount
that was an important step I left out.

For bastion-host IaC updated rhel bh userdata
to install eksctl.