Jenkinsfile, web-app deployment and nlb service:
Updated the Jenkinsfile in regards to the iam service account I've been
trying to use to create an AWS Load Balancer Controller. It seems
the ias serviceaccount was missing a role parameter that gave it its
permissions. Added the Role parameter.

Am also trying to deploy an nlb instead of an alb because it only
requires a service, and not a service and an ingress.

Commented out serviceAccountName from pod spec on deployment because
the example I was following didn't have it.