IaC - EKS VPC - Added a second Public Subnet:
The error with the load balancers were they require
two public subnets to load balance between, and I only had 1 public
subnet and 2 private subnets. This is because this cluster was initially
going to be private. The cluster now has 2 public and 2 private subnets.

I doubt this will work the first go around.