Kubernetes deployment - web-app-nlb:
Removed annotation that was going to list the
subnets because I tagged the subnets with
kubernetes tags, that should be all that is required.

In last comment I commented out all the crap I added to swap out the
subnet variable placeholders with the values of the subnet it. I hope
to be able to delete if only tagging the subnets works.