Major update - I should have made smaller commits:
Jenkinsfile/IaC:
    - Added variable for:
        KUBE_VERSION
        KUBE_NAMESPACE
        KUBE_LOAD_BALANCER_TYPE

Most of the updates were simple, replacing existing values
with the Variable name. For files other than Jenkinsfile
I added a command in Jenkinsfile to swap the variable for the value.

For KUBE_LOAD_BALANCER_TYPE, I added logic at the end of Jenkinsfile
to deploy either the NLB or ALB based on this value. It's only a
simple if/elseif/else.

I'm sure troubleshooting will be required, but the changes make
the pipeline more re-usable and resilient to change.