Jenkinsfile - IaC:
For Jenkinsfile I only updated the service name for the nlb service,
as it was mispelled. And I modified the variable names for SERVICE_NAME
to KUBE_NAMESPACE.

Also, authenticating the Bastion Host to the cluster data won't work,
unless I make the UserData sleep for 5 minutes because the bastion_host
instance role hasn't been trusted by the cluster yet.