Fully Working - Merging to main:
Everything is fully working, so I'm merging to main.
This commit added back in configuring the BH kubectl
to connect to the EKS Cluster. I just added a 3 min sleep in the
UserData to give the deployment time to trust the BH Instance Role.
It probably could be shortened to 1 minute tops, but figured better
safe than sorry.