Jenkinsfile - Cleaned up syntax, mostly cosmetic:
Cleaned up the Jenkinsfile with mostly cosmetic
changes. Only functional change was to reduce sleep
and attempt to fix the wait logic for the creation of
the AWS Load Balancer Controller.

I changed the IaC for deploying the EKS VPC back to
a hardcoded cidrRange because ever since I used the
parameter, a 2nd empty EKS VPC has been getting created.
Not sure if this is a bug but it started after changing
that value to a parameter. Seeing if this resolves the
issue.