IaC - Bastion Host IAM:
Updating IAM Policy for Bastion Host since that Role will now be
launching the EKS Cluster, and performing other actions. Will have to
reduce permissions later.

There is an issue with the Bastion Host connecting to the EKS Cluster. I
think the issue is with the peering connection. Launching the
public-cluster to verify how the peering connection looks.