README - Updated for this branch:
Updated README to explain how this branch is different because it uses a
Private API Endpoint.

Added 10 minute sleep command to eks-deploy-script.sh after the cluster
is created because there is no other reason I can figure out why the
script can't connect to the EKS Cluster with kubectl, but I can run the
same script when I log on and connect to it.