IaC - EKS Nodes connecting to Cluster:
The EKS Nodes stil aren't connecting to the cluster. I fixed the S3
endpoint to actually be a gateway. Other than that everything else seems
to match up with what the docs say to do. I added the EKS Keypair back
in so I can log into the EKS Node to troubleshoot.