kubernetes & pipeline - added a bunch of features:
Added a bunch of features the I added to the private clusters. Mainly
variables for deploying whatever app the user specifies to the cluster,
and building a repo and pushing the image to it.

There is a stage where an image can be built, but I just push an image
from docker hub.